Add in forward speed of the storm

Adds the storm's forward speed component (i.e., motion asymmetry) back into the estimated surface wind speed at a grid point location after rotational winds have been modeled for the location.

add_forward_speed(wind_sfc_sym, tcspd_u, tcspd_v, swd, cdist, Rmax)

Arguments

wind_sfc_sym

A numeric value with the estimated symmetric surface wind speed at the grid point, in meters / second

tcspd_u

A numeric value with the u-component of the tropical cyclone speed in meters / second

tcspd_v

A numeric value with the v-component of the tropical cyclone speed in meters / second

swd

A numeric value with the surface wind direction in degrees

cdist

A numeric value with the radius from the storm's center to the grid point in kilometers

Rmax

A numeric value with the radius at which maximum winds occur in kilometers

Value

windspeed A numeric value with the asymmetric surface windspeed at the location, in meters/second

Details

This function uses equation 12 from Phadke et al. (2003).

References

Phadke AC, Martino CD, Cheung KF, and Houston SH. 2003. Modeling of tropical cyclone winds and waves for emergency management. Ocean Engineering 30(4):553-578.