Add inflow angle

This function adds an inflow angle to the angle of the wind direction. It calculates an inflow angle as a function of the distance from the storm center to a location (Phadke et al. 2003), and then adds 20 degrees to this inflow angle to account for the location being over land rather than over water.

add_inflow(gwd, cdist, Rmax, tclat)

Arguments

gwd

A numeric vector giving direction of gradient wind at a location, in degrees. Due east is 0 degrees, due north 90 degrees, etc.

cdist

A numeric vector giving radius (in kilometers) from the storm center to the location being modeled.

Value

Numeric vector with the gradient wind direction (in degrees), adjusted with an inflow angle appropriate for being over land and for the location's distance from the storm's center.

Details

This function uses equations 11a-c from Phadke et al. (2003).

Note

This function is only appropriate for modeling wind speeds for locations that are over land.

References

Phadke AC, Martino CD, Cheung KF, and Houston SH. 2003. Modeling of tropical cyclone winds and waves for emergency management. Ocean Engineering 30(4):553-578.

Examples

add_inflow(gwd = 160, cdist = 100, Rmax = 20, tclat = 32)
#> [1] 205

add_inflow(gwd = 160, cdist = 100, Rmax = 20, tclat = -32)
#> [1] 115