Add Willoughby inputs and parameters

Calculates a number of inputs and parameters needed for the Willoughby model for each observation of a storm track. This function intakes an imputed storm track and calculates these parameters for each observation in the storm tracks.These inputs and parameters are later used to model wind speed at each grid point for every observation.

add_wind_radii(full_track = create_full_track())

Arguments

full_track

A dataframe with interpolated hurricane track data, as created by create_full_track. This dataframe must have the following columns:

date: Date-time, in POSIXct format and UTC time zone

tclat: Latitude (decimal degree)

tclon: Longitude (decimal degrees), expressed as positive values (this model assumes all longitudes will be in the Western hemisphere)

vmax: Maximum 10-meter sustained wind speed (m / s)

Value

The input dataset, but with columns added for the Willoughby parameters for every observations. Added columns include:

tcdir: The bearing of the storm, in degrees, with 0 degrees indicating the storm is moving due east, 90 degrees indicating the storm is moving due north, etc.

tcspd_u, tcspd_v: The u- and v-components of the forward speed of the storm, in meters per second

vmax_gl: The maximum gradient-level 1-minute sustained wind, in m / s

Rmax: The radius to maximum winds, in kilometers

X1, n, A, R1, R2: Parameters needed for the Willoughby wind model

References

Willoughby HE, Darling RWR, and Rahn ME. 2006. Parametric representation of the primary hurricane vortex. Part II: A new family of sectionally continuous profiles. Monthly Weather Review 134(4):1102-1120.

Examples

if (FALSE) {
data("floyd_tracks")
full_track <- create_full_track(hurr_track = floyd_tracks)
with_wind_radii <- add_wind_radii(full_track = full_track)
head(with_wind_radii)
}