Calculate bearing from one location to another

Calculates the bearing of a second location, as seen from the first location, based on latitude and longitude coordinates for both locations.

calc_bearing(tclat_1, tclon_1, tclat_2, tclon_2)

Value

A numeric vector giving the direction of the second location from the first location, in degrees. A direction of 0 degrees indicates the second location is due east of the first, 90 degrees indicates the second location is due north of the first, etc (i.e., polar, rather than meteorological, coordinate system).

Details

This function uses the following equations to calculate the bearing from one latitude-longitude pair to another:

$$S = cos(\phi_2) * sin(L_2 - L_1)$$

$$C = cos(\phi_1) * sin(\phi_2) - sin(\phi_1) * cos(\phi_2) * cos(L_2 - L_1)$$

$$\theta = 90 - atan2(S, C) * \frac{180}{\pi}$$

where:

\(\phi_1\): Latitude of first location, in radians

\(L_1\): Longitude of first location, in radians

\(\phi_2\): Latitude of second location, in radians

\(L_2\): Longitude of second location, in radians

\(S, C\): Intermediary results

\(\theta\): Direction of the storm movement, in degrees

In cases where this equation results in values below 0 degrees or above 360 degrees, the function applies modular arithmetic to bring the value back within the 0--360 degree range.