Calculate wind speed at grid points

Uses the Willoughby wind model (Willoughby et al. 2006) to calculate wind speed at a grid point location. This function calculates a wind time series at just one location.

calc_grid_wind(
  grid_point = stormwindmodel::county_points[1, ],
  with_wind_radii = add_wind_radii(),
  max_dist = 2222.4
)

Arguments

grid_point

A one-row dataframe with the grid id, latitude, and longitude for a single location for which you want to model winds.

with_wind_radii

A dataframe of storm tracks, including inputs and parameters for the Willoughby wind model, as created by add_wind_radii.

max_dist

A numeric value giving (in kilometers) the maximum distance from the storm's center to model storm-associated winds. Any value beyond this distance will be automatically set to 0 m / s for storm-associated winds. The default value is 2222.4 km (1200 nautical miles).

Value

A dataframe with date (date) and modeled wind speed (windspeed, in m / s) at the grid point location for all storm observations.

References

Knaff JA, DeMaria M, Molenar DA, Sampson CR, and Seybold MG. 2011. An automated, objective, multiple-satellite-platform tropical cyclone surface wind speed analysis. Journal of Applied Meteorology and Climatology 50(10):2149-2166

Phadke AC, Martino CD, Cheung KF, and Houston SH. 2003. Modeling of tropical cyclone winds and waves for emergency management. Ocean Engineering 30(4):553-578.

Willoughby HE, Darling RWR, and Rahn ME. 2006. Parametric representation of the primary hurricane vortex. Part II: A new family of sectionally continuous profiles. Monthly Weather Review 134(4):1102-1120.

Examples

if (FALSE) {
data("floyd_tracks")
data("county_points")
full_track <- create_full_track(hurr_track = floyd_tracks)
with_wind_radii <- add_wind_radii(full_track = full_track)
wind_grid <- calc_grid_wind(grid_point = county_points[1, ],
                            with_wind_radii = with_wind_radii)
head(wind_grid)
}