Calculate gradient wind direction at a point

This function determines the bearing from the storm's center to a grid point location and then calculates the direction of gradient winds at that grid point. Gradient winds will be perpendicular to the direction from the storm's center to the grid point. In the Northern hemisphere, cyclonic winds are counterclockwise, so you add 90 degrees to the bearing from the storm center to the grid point when using polar coordinates. In the Southern hemisphere, cyclonic winds are clockwise, so you subtract 90 degrees.

Usage

calc_gwd(tclat, tclon, glat, glon)

Arguments

tclat

A numeric value with the latitude of the tropical cyclone's center in radians

tclon

A numeric value with the longitude of the tropical cyclone's center in radians

glat

A numeric value with the latitude of the grid point in radians

glon

A numeric value with the longitude of the grid point in radians

Value

A numeric value with the the direction of gradient storm winds at a location, in polar coordinates (i.e., due East is 0 degrees, due North is 90 degrees, etc.)