Hurricane Floyd tracks data

A dataframe containing hurricane best tracks for Hurricane Floyd in 1999. This dataframe is included for use as an example hurricane tracks dataframe in the package documentation. This data originally came from the Extended Best Track dataset: http://rammb.cira.colostate.edu/research/tropical_cyclones/tc_extended_best_track_dataset/

Usage

floyd_tracks

Format

A dataframe with 48 rows and 4 variables:

date

A character string giving the date and time of the observation

latitude

A numeric vector giving the storm's latitude at that observation time

longitude

A numeric vector giving the storm's longitude at that observation time

wind

A numeric vector giving the estimated maximum sustained wind of that storm at that observation time, in knots

Source

http://rammb.cira.colostate.edu/research/tropical_cyclones/tc_extended_best_track_dataset/