Numerically solve Willoughby Eqn. 3 for xi

This function uses the Newton-Raphson method to solve equation 3 (the dual-exponential profile version) for \(\xi\) in Willoughby et al. (2006). This value of \(\xi\) can then be used to determine \(R_1\) for that storm observation.

Usage

solve_for_xi(xi0 = 0.5, eq3_right, eps = 0.001, itmax = 100)

Arguments

xi0

A numeric value giving the starting guess for \(\xi\)

eq3_right

A numerical value with the right-hand side of Willoughby et al. (2006), Eqn. 3, the dual-exponential version. This value is calculated at each storm observation point using the will3_right function, defined in C++ code for this package.

eps

The convergence threshold for determining if the algorithm has converged.

itmax

The maximum number of iterations to try before deciding that the algorithm did not converge.

Note

If this algorithm does not converge, the function returns a missing value for \(\xi\).

References

Jones O, Maillardet R, and Robinson A. 2009. Introduction to Scientific Programming and Simulation Using R. Boca Raton, FL: Chapman & Hall/CRC Press.

Press WH, Teukolsky SA, Vetterling WT, and Flannery BP. 2002. Numerical Recipes in C++: The Art of Scientific Computing. 2nd ed. Cambridge, UK: Cambridge University Press.

Willoughby HE, Darling RWR, and Rahn ME. 2006. Parametric representation of the primary hurricane vortex. Part II: A new family of sectionally continuous profiles. Monthly Weather Review 134(4):1102-1120.