Model wind speed at a grid point for a storm track observation

Models the gradient wind speed at a certain radius from a storm's center. To do this, it uses different equations and subfunctions depending on how large the radius is (see details). This function requires, as inputs, Willoughby wind model parameters calculated using the add_wind_radii function.

Usage

will1(cdist, Rmax, R1, R2, vmax_gl, n, A, X1, X2 = 25)

Arguments

cdist

A numeric value with the distance between the grid point and the center of the storm (in km)

Rmax

A numeric value with the radius at which the maximum wind occurs (in km)

R1

A numeric value with the lower boundary of the transition zone (in km from the storm center)

R2

A numeric value with the upper boundary of the transition zone (in km from the storm center)

vmax_gl

A numeric value with the maximum gradient level 1-minute sustained wind, in (m/s)

n

A numeric value giving the power by which the wind is assumed to increase with radius within the center of the storm (i.e., from the center to the radius of maximum wind)

A

A numeric value that is a parameter between 0 and 1 for the Willoughby model

X1

A numeric value that is a parameter for the Willoughby model

X2

A numeric value as a parameter for the Willoughby model, set to 25 (Willoughby, Darling, and Rahn 2006)

Value

Returns a numeric vector with gradient wind speed at a radius of \(r\) from the storm's center, in meters per second.

Details

If \(r \le R_1\), this function is calculating the equation:

$$V(r) = V_i = V_{max} \left( \frac{r}{R_{max}} \right)^n$$

where:

\(V(r)\): Maximum sustained gradient wind speed at a radius of \(r\) from the storm's center

\(r\): Radius from the storm center, in kilometers

\(V_{max,G}\): Maximum sustained gradient wind speed of the storm, in meters per second

\(R_1\): A parameter for the Willoughby wind model (radius to start of transition region)

\(R_{max}\): Radius (in kilometers) to highest winds

\(n\): A parameter for the Willoughby wind model

If \(R_2 < r\), this function is calculating the equation:

$$V(r) = V_o = V_{max}\left[(1 - A) e^\frac{R_{max} - r}{X_1} + A e^\frac{R_{max} - r}{X_2}\right]$$

where:

\(V(r)\): Maximum sustained gradient wind speed at a radius of \(r\) kilometers from the storm's center

\(r\): Radius from the storm center, in kilometers

\(V_{max,G}\): Maximum sustained gradient wind speed of the storm, in meters per second

\(R_{max}\): Radius (in kilometers) to highest winds

\(A\), \(X_1\), \(X_2\): Parameters for the Willoughby wind model

If \(R_1 < r \le R_2\), this function uses the equations:

$$\xi = \frac{r - R_1}{R_2 - R_1}$$

and, if \(0 \le \xi < \le 1\) (otherwise, \(w = 0\)):

$$w = 126 \xi^5 - 420 \xi^6 + 540 \xi^7- 315 \xi^8 + 70 \xi^9$$

and then:

$$V(r) = V_i (1 - w) + V_o w, (R_1 \le r \le R_2)$$

where, for this series of equations:

\(V(r)\): Maximum sustained gradient wind speed at a radius of \(r\) kilometers from the storm's center

\(r\): Radius from the storm center, in kilometers

\(w\): Weighting variable

\(R_1\), \(R_2\): Parameters for the Willoughby wind model

References

Willoughby HE, Darling RWR, and Rahn ME. 2006. Parametric representation of the primary hurricane vortex. Part II: A new family of sectionally continuous profiles. Monthly Weather Review 134(4):1102-1120.