Calculate the function value and derivative for Willoughby Eqn. 3

Calculates values of both the function and the derivative of the function for which you are trying to solve the root in Eqn. 3 (the version using the dual exponential profile) of Willoughby et al. (2006).

will3_deriv_func(xi, eq3_right)

Arguments

xi

A numerical value for \(\xi\) from Willoughby et al. (2006), Eqn. 2

eq3_right

A numerical value with the right-hand side of Willoughby et al. (2006), Eqn. 3, the dual-exponential version. This value is calculated at each storm observation point using the will3_right function.

Value

A numeric vector of length two. The first value is the calculated value of \(f'(x)\) for \(x = \xi\), while the second value is the calculated value of \(f(x)\) for \(x = \xi\). These two values are used in iterating through the Newton-Raphson method to determine \(\xi\).

References

Willoughby HE, Darling RWR, and Rahn ME. 2006. Parametric representation of the primary hurricane vortex. Part II: A new family of sectionally continuous profiles. Monthly Weather Review 134(4):1102-1120.