Calculate radius of maximum winds

Calculates the radius at which the maximum wind occurs (\(R_{max}\)), based on the maximum gradient-level 1-min sustained wind (\(V_{max,G}\)) and latitude (\(\phi\)). This function implements Willoughby et al. (2006), Equation 7a.

Usage

will7a(vmax_gl, tclat)

Arguments

vmax_gl

Maximum gradient-level 1-min sustained wind for the tropical cyclone

tclat

Numeric vector of the value of latitude, in degrees.

Value

A numeric vector with \(R_{max}\), the radius of maximum winds, in kilometers.

Details

This function fits the following equation: $$R_{max} = 46.4 e^{- 0.0155 V_{max,G} + 0.0169\phi}$$ where:

\(R_{max}\): Radius from the storm center to the point at which the maximum wind occurs (km)

\(V_{max,G}\): Tangential wind component of the gradient-level maximum wind speed (m / s)

\(\phi\): Absolute value of latitude (in decimal degrees)

References

Willoughby HE, Darling RWR, and Rahn ME. 2006. Parametric representation of the primary hurricane vortex. Part II: A new family of sectionally continuous profiles. Monthly Weather Review 134(4):1102-1120.