Add Willoughby inputs and parameters

Calculates a number of inputs and parameters needed for the Willoughby model for each observation of a storm track. This function intakes an imputed storm track and calculates these parameters for each observation in the storm tracks.These inputs and parameters are later used to model wind speed at each grid point for every observation.

Usage

add_wind_radii(full_track = create_full_track())

Arguments

full_track

A dataframe with interpolated hurricane track data, as created by create_full_track. This dataframe must have the following columns:

date: Date-time, in POSIXct format and UTC time zone

tclat: Latitude (decimal degree)

tclon: Longitude (decimal degrees), expressed as positive values (this model assumes all longitudes will be in the Western hemisphere)

vmax: Maximum 10-meter sustained wind speed (m / s)

Value

The input dataset, but with columns added for the Willoughby parameters for every observations. Added columns include:

tcdir: The bearing of the storm, in degrees, with 0 degrees indicating the storm is moving due east, 90 degrees indicating the storm is moving due north, etc.

tcspd_u, tcspd_v: The u- and v-components of the forward speed of the storm, in meters per second

vmax_gl: The maximum gradient-level 1-minute sustained wind, in m / s

Rmax: The radius to maximum winds, in kilometers

X1, n, A, R1, R2: Parameters needed for the Willoughby wind model

References

Willoughby HE, Darling RWR, and Rahn ME. 2006. Parametric representation of the primary hurricane vortex. Part II: A new family of sectionally continuous profiles. Monthly Weather Review 134(4):1102-1120.

Examples

data("floyd_tracks")
full_track <- create_full_track(hurr_track = floyd_tracks)
with_wind_radii <- add_wind_radii(full_track = full_track)
head(with_wind_radii)
#>                  date    tclat     tclon    tcdir   tcspd_u  tcspd_v  vmax_gl
#> 1 1999-09-07 18:00:00 14.60000 -45.60000 160.7715 -6.510216 2.270724 10.45952
#> 2 1999-09-07 18:15:00 14.61835 -45.65439 160.7863 -6.509378 2.268546 10.57943
#> 3 1999-09-07 18:30:00 14.63669 -45.70879 160.8174 -6.508274 2.264199 10.69988
#> 4 1999-09-07 18:45:00 14.65499 -45.76317 160.8649 -6.506877 2.257674 10.82088
#> 5 1999-09-07 19:00:00 14.67323 -45.81756 160.9287 -6.505188 2.248970 10.94242
#> 6 1999-09-07 19:15:00 14.69141 -45.87193 161.0090 -6.503234 2.238097 11.06449
#>       Rmax       X1         n          A       R1       R2
#> 1 50.49715 323.8680 0.5018371 0.02741166 35.19583 60.19583
#> 2 50.41901 323.6602 0.5034941 0.02788177 35.10915 60.10915
#> 3 50.34057 323.4513 0.5051589 0.02835462 35.02214 60.02214
#> 4 50.26179 323.2412 0.5068317 0.02883039 34.93479 59.93479
#> 5 50.18266 323.0299 0.5085126 0.02930918 34.84708 59.84708
#> 6 50.10319 322.8174 0.5102013 0.02979099 34.75901 59.75901