Determine hurricane winds at locations

This function inputs a storm track and a dataset of locations and calculates highest wind speeds (sustained and maximum) and duration of winds above a certain speed at each location. It is assumed that track data entered has been measured in knots, at 10 m above surface level, and with 1-minute averaging period. The dataset of locations can either be a regularly-spaced grid or can be the central points of locations, like counties or cities. For counties in the eastern half of the United States, the county_points dataset that comes with the package can be used as the grid_point input.

Usage

get_grid_winds(
  hurr_track = stormwindmodel::floyd_tracks,
  grid_df = stormwindmodel::county_points,
  tint = 0.25,
  gust_duration_cut = 20,
  sust_duration_cut = 20,
  max_dist = 2222.4
)

Arguments

hurr_track

Dataframe with hurricane track data for a single storm. The dataframe must include columns for date-time (year, month, day, hour, minute; e.g., "198808051800" for August 5, 1988, 18:00 UTC), latitude, longitude (in degrees East), and wind speed (in knots). The column names for each of these must be date, latitude, longitude, and wind. See the example floyd_tracks dataset for an example of the required format.

grid_df

A dataframe of locations at which to calculate wind characteristics. This dataframe must include columns for latitude and longitude for each point, and these columns must be named glat and glon. The latitudes and longitudes should be in decimal degrees, with longitudes being entered in degrees East. Therefore Western hemisphere (so, almost all those for Atlantic basin storms) should be expressed as negative values.

tint

Interval (in hours) to which to interpolate the tracks. The default is 0.25 (i.e., 15 minutes).

gust_duration_cut

The wind speed, in meters per second, to use as a cutoff point for determining the duration of gust winds. The function will calculate the minutes during the storm when surface-level gust winds were above this speed at the location

sust_duration_cut

The wind speed, in meters per second, to use as a cutoff point for determining the duration of gust winds. The function will calculate the minutes during the storm when surface-level gust winds were above this speed at the location.

max_dist

A numeric value giving (in kilometers) the maximum distance from the storm's center to model storm-associated winds. Any value beyond this distance will be automatically set to 0 m / s for storm-associated winds. The default value is 2222.4 km (1200 nautical miles).

Value

The dataframe of locations input, with the following columns of wind characteristics added for each location:

vmax_gust: Maximum value of surface-level (10 meters) gust winds, in meters per second, over the length of the storm at the given location

vmax_sust: Maximum value of surface-level (10 meters) sustained winds, in meters per second, over the length of the storm at the given location

gust_duration: Length of time, in minutes, that surface-level gust winds were above a specified value (default is 20 m / s)

sust_duration: Length of time, in minutes, that surface-level sustained winds were above a specified value (default is 20 m / s)

Note

This function can take a few minutes to run, depending on the number of locations that are being modeled.

data("floyd_tracks") data("county_points") grid_winds <- get_grid_winds(hurr_track = floyd_tracks, grid_df = county_points)